requirements.txt
Django==2.2.6
djangorestframework==3.9.3
requests==2.22.0
github3.py==1.0.2
